NB class object

The functions NB returns an object of the type NB.

An object of class 'NB' is a list with at least the following components:

prior

Matrix with prior probabilities, i.e. marginal class probabilities Pr(class)

pcond

list of matrices with conditional probabilities of the features given the classes (one list element per class), Pr(x|class)

classes

Names (character vector) of the classes

xvar

number of repetitions of the CV

xmodel

Number of folds of the CV

model

Number of folds of the CV

Value

objects of the S3 class 'NB'

S3 generics

The following S3 generic functions are available for an object of class NB:

predictPredict class probabilities for new features data.

printBasic print method.

See also

Examples

## See example(NB) for examples