AIPW estimator for Average Causal Effect

Augmented Inverse Probability Weighting estimator for the Average Causal Treatment Effect.

ace(
  formula,
  data,
  weights,
  binary = TRUE,
  nuisance = NULL,
  propensity = nuisance,
  all,
  missing = FALSE,
  labels = NULL,
  ...
)

Arguments

formula
Formula (see details below)

data
data.frame

weights
optional frequency weights

binary
Binary response (default TRUE)

nuisance
outcome regression formula

propensity
propensity model formula

all
If TRUE all standard errors are calculated (default TRUE when exposure only has two levels)

missing
If TRUE a missing data (AIPW) estimator is returned

labels
Optional treatment labels

...
Additional arguments to lower level functions

Value

An object of class 'ace.targeted' is returned. See targeted-class for more details about this class and its generic functions.

Details

The formula may either be specified as: response ~ treatment | nuisance-formula | propensity-formula

For example: ace(y~a | x+z | x*z, data=...)

Alternatively, as a list: ace(list(y~a, ~x+z, ~x*z), data=...)

Or using the nuisance (and propensity argument): ace(y~a, nuisance=~x+z, ...)

Examples

m <- lvm(y ~ a+x, a~x)
distribution(m,~ a+y) <- binomial.lvm()
d <- sim(m,1e3,seed=1)

a <- ace(y ~ a, nuisance=~x, data=d)
summary(a)

#> 
#> Augmented Inverse Probability Weighting estimator
#>   Response y (Outcome model: logistic regression):
#> 	 y ~ x 
#>   Exposure a (Propensity model: logistic regression):
#> 	 a ~ x 
#> 
#>                   Estimate Std.Err    2.5%   97.5%    P-value
#>  a=0               0.48506 0.02626  0.4336  0.5365  3.458e-76
#>  a=1               0.67794 0.02225  0.6343  0.7215 6.005e-204
#> Outcome model:                                               
#>  (Intercept)       0.44427 0.07306  0.3011  0.5875  1.196e-09
#>  x                 1.06929 0.08537  0.9020  1.2366  5.408e-36
#> Propensity model:                                            
#>  (Intercept)       0.06214 0.09258 -0.1193  0.2436  5.021e-01
#>  x                -0.92905 0.15311 -1.2291 -0.6289  1.297e-09
#> 
#> Average Causal Effect (constrast: 'a=0' vs. 'a=1'):
#> 
#>    Estimate Std.Err    2.5%   97.5%   P-value
#> RR   0.7155 0.04356  0.6301  0.8009 1.259e-60
#> OR   0.4475 0.06268  0.3246  0.5703 9.383e-13
#> RD  -0.1929 0.03295 -0.2575 -0.1283 4.791e-09
#> 

# Multiple treatments
m <- lvm(y ~ a+x, a~x)
distribution(m,~ y) <- binomial.lvm()
m <- ordinal(m, K=4, ~a)
transform(m, ~a) <- factor
d <- sim(m,1e4)
(a <- ace(y~a|a*x|x, data=d))

#>     Estimate  Std.Err   2.5%  97.5%    P-value
#> a=0   0.2324 0.019002 0.1951 0.2696  2.196e-34
#> a=1   0.3352 0.018523 0.2989 0.3715  3.416e-73
#> a=2   0.4162 0.015837 0.3852 0.4472 3.236e-152
#> a=3   0.5967 0.008458 0.5801 0.6132  0.000e+00

# Comparison with randomized experiment
m0 <- cancel(m, a~x)
d0 <- sim(m0,2e5)
lm(y~a-1,d0)

#> 
#> Call:
#> lm(formula = y ~ a - 1, data = d0)
#> 
#> Coefficients:
#>     a0      a1      a2      a3  
#> 0.2243  0.3386  0.4129  0.6195  
#> 

# Choosing a different contrast for the association measures
summary(a, contrast=c(2,4))

#> 
#> Augmented Inverse Probability Weighting estimator
#>   Response y (Outcome model: logistic regression):
#> 	 y ~ a * x 
#>   Exposure a (Propensity model: logistic regression):
#> 	 a ~ x 
#> 
#>     Estimate  Std.Err   2.5%  97.5%    P-value
#> a=0   0.2324 0.019002 0.1951 0.2696  2.196e-34
#> a=1   0.3352 0.018523 0.2989 0.3715  3.416e-73
#> a=2   0.4162 0.015837 0.3852 0.4472 3.236e-152
#> a=3   0.5967 0.008458 0.5801 0.6132  0.000e+00
#> 
#> Average Causal Effect (constrast: 'a=1' vs. 'a=3'):
#> 
#>    Estimate Std.Err    2.5%   97.5%   P-value
#> RR   0.5618 0.03211  0.4989  0.6247 1.505e-68
#> OR   0.3408 0.03084  0.2804  0.4013 2.169e-28
#> RD  -0.2615 0.02042 -0.3015 -0.2214 1.560e-37
#>