Superlearner (stacked/ensemble learner)

This function creates a predictor object (class ml_model) from a list of existing ml_model objects. When estimating this model a stacked prediction will be created by weighting together the predictions of each of the initial models. The weights are learned using cross-validation.

predictor_sl(
  model.list,
  info = NULL,
  nfolds = 5L,
  meta.learner = NULL,
  model.score = mse,
  ...
)

Arguments

model.list

List of ml_model objects (i.e. predictor_glm)

info

Optional model description to store in model object

nfolds

Number of folds to use in cross validation

meta.learner

meta.learner function (default non-negative least squares). Must be a function of the response (nx1 vector), y, and the predictions (nxp matrix), pred.

model.score

model scoring method (see ml_model)

...

additional argument to superlearner

References

Luedtke & van der Laan (2016) Super-Learning of an Optimal Dynamic Treatment Rule, The International Journal of Biostatistics.

See also