AIPW estimator for Average Treatement Effect

Augmented Inverse Probability Weighting estimator for the Average (Causal) Treatment Effect.

ate(
  formula,
  data = parent.frame(),
  weights,
  binary = TRUE,
  nuisance = NULL,
  propensity = nuisance,
  all,
  missing = FALSE,
  labels = NULL,
  ...
)

Arguments

formula

Formula (see details below)

data

data.frame

weights

optional frequency weights

binary

Binary response (default TRUE)

nuisance

outcome regression formula

propensity

propensity model formula

all

If TRUE all standard errors are calculated (default TRUE when exposure only has two levels)

missing

If TRUE a missing data (AIPW) estimator is returned

labels

Optional treatment labels

...

Additional arguments to lower level functions

Value

An object of class 'ate.targeted' is returned. See targeted-class

for more details about this class and its generic functions.

Details

The formula may either be specified as: response ~ treatment | nuisance-formula | propensity-formula

For example: ate(y~a | x+z+a | x*z, data=...)

Alternatively, as a list: ate(list(y~a, ~x+z, ~x*z), data=...)

Or using the nuisance (and propensity argument): ate(y~a, nuisance=~x+z, ...)

Author

Klaus K. Holst

Examples

m <- lvm(y ~ a+x, a~x)
distribution(m,~ a+y) <- binomial.lvm()
d <- sim(m,1e3,seed=1)

a <- ate(y ~ a, nuisance=~x, data=d)
summary(a)
#> 
#> Augmented Inverse Probability Weighting estimator
#>   Response y (Outcome model: logistic regression):
#> 	 y ~ x + a 
#>   Exposure a (Propensity model: logistic regression):
#> 	 a ~ x 
#> 
#>                   Estimate Std.Err     2.5%   97.5%    P-value
#>  a=0               0.48496 0.02311  0.43967  0.5302  8.607e-98
#>  a=1               0.67788 0.02240  0.63397  0.7218 4.249e-201
#> Outcome model:                                                
#>  (Intercept)      -0.02229 0.10163 -0.22149  0.1769  8.264e-01
#>  x                 0.90226 0.08857  0.72866  1.0759  2.276e-24
#>  a                 1.01052 0.15195  0.71270  1.3083  2.926e-11
#> Propensity model:                                             
#>  (Intercept)       0.06214 0.06914 -0.07337  0.1977  3.688e-01
#>  x                -0.92905 0.08035 -1.08653 -0.7716  6.408e-31
#> 
#> Average Treatment Effect (constrast: 'a=0' vs. 'a=1'):
#> 
#>    Estimate Std.Err    2.5%   97.5%   P-value
#> RR   0.7154 0.04004  0.6369  0.7939 2.114e-71
#> OR   0.4474 0.05952  0.3308  0.5641 5.611e-14
#> RD  -0.1929 0.03099 -0.2537 -0.1322 4.785e-10
#> 

# Multiple treatments
m <- lvm(y ~ a+x, a~x)
distribution(m,~ y) <- binomial.lvm()
m <- ordinal(m, K=4, ~a)
transform(m, ~a) <- factor
d <- sim(m, 1e4, seed=1)
(a <- ate(y~a|a*x|x, data=d))
#>     Estimate  Std.Err   2.5%  97.5%    P-value
#> a=0   0.2406 0.018291 0.2047 0.2764  1.638e-39
#> a=1   0.3638 0.017241 0.3300 0.3976  7.816e-99
#> a=2   0.4003 0.015525 0.3699 0.4307 1.286e-146
#> a=3   0.6054 0.007469 0.5908 0.6201  0.000e+00

# Comparison with randomized experiment
m0 <- cancel(m, a~x)
d0 <- sim(m0,2e5)
lm(y~a-1,d0)
#> 
#> Call:
#> lm(formula = y ~ a - 1, data = d0)
#> 
#> Coefficients:
#>     a0      a1      a2      a3  
#> 0.2184  0.3425  0.4083  0.6200  
#> 

# Choosing a different contrast for the association measures
summary(a, contrast=c(2,4))
#> 
#> Augmented Inverse Probability Weighting estimator
#>   Response y (Outcome model: logistic regression):
#> 	 y ~ a * x 
#>   Exposure a (Propensity model: logistic regression):
#> 	 a ~ x 
#> 
#>     Estimate  Std.Err   2.5%  97.5%    P-value
#> a=0   0.2406 0.018291 0.2047 0.2764  1.638e-39
#> a=1   0.3638 0.017241 0.3300 0.3976  7.816e-99
#> a=2   0.4003 0.015525 0.3699 0.4307 1.286e-146
#> a=3   0.6054 0.007469 0.5908 0.6201  0.000e+00
#> 
#> Average Treatment Effect (constrast: 'a=1' vs. 'a=3'):
#> 
#>    Estimate Std.Err    2.5%   97.5%   P-value
#> RR   0.6009 0.02911  0.5438  0.6579 1.240e-94
#> OR   0.3726 0.02964  0.3146  0.4307 2.964e-36
#> RD  -0.2416 0.01849 -0.2779 -0.2054 4.999e-39
#>