Targeted inference

Methods for targeted and semiparametric inference including augmented inverse probability weighted estimators for missing data and causal inference.

Author

Klaus K. Holst Maintainer: <klaus@holst.it>

Examples

example(riskreg)

#> 
#> riskrg> m <- lvm(a[-2] ~ x,
#> riskrg+         lp.target[1] ~ 1,
#> riskrg+         lp.nuisance[-1] ~ 2*x)
#> 
#> riskrg> distribution(m,~a) <- binomial.lvm("logit")
#> 
#> riskrg> m <- binomial.rr(m, "y","a","lp.target","lp.nuisance")
#> 
#> riskrg> d <- sim(m,5e2,seed=1)
#> 
#> riskrg> I <- model.matrix(~1, d)
#> 
#> riskrg> X <- model.matrix(~1+x, d)
#> 
#> riskrg> with(d, riskreg_mle(y, a, I, X, type="rr"))
#>    Estimate Std.Err    2.5%   97.5%   P-value
#> p1   0.9510  0.1176  0.7205  1.1815 6.196e-16
#> p2  -0.8606  0.2973 -1.4433 -0.2779 3.795e-03
#> p3   2.2874  0.3844  1.5339  3.0409 2.679e-09
#> 
#> riskrg> with(d, riskreg_fit(y, a, nuisance=X, propensity=I, type="rr"))
#>    Estimate Std.Err   2.5% 97.5%   P-value
#> p1   0.9528  0.1165 0.7246 1.181 2.791e-16
#> 
#> riskrg> riskreg(y ~ a | 1 | x ,  data=d, type="rr")
#>             Estimate Std.Err   2.5% 97.5%  P-value
#> (Intercept)   0.9397  0.1193 0.7058 1.174 3.38e-15
#> 
#> riskrg> ## Model with same design matrix for nuisance and propensity model:
#> riskrg> with(d, riskreg_fit(y, a, nuisance=X, type="rr"))
#>    Estimate Std.Err   2.5% 97.5%  P-value
#> p1   0.9397  0.1193 0.7058 1.174 3.38e-15
#> 
#> riskrg> a <- riskreg(y ~ a, nuisance=~x,  data=d, type="rr")
#> 
#> riskrg> a
#>             Estimate Std.Err   2.5% 97.5%  P-value
#> (Intercept)   0.9397  0.1193 0.7058 1.174 3.38e-15

example(ace)

#> 
#> ace> m <- lvm(y ~ a+x, a~x)
#> 
#> ace> distribution(m,~ a+y) <- binomial.lvm()
#> 
#> ace> d <- sim(m,1e3,seed=1)
#> 
#> ace> a <- ace(y ~ a, nuisance=~x, data=d)
#> 
#> ace> summary(a)
#> 
#> Augmented Inverse Probability Weighting estimator
#>   Response y (Outcome model: logistic regression):
#> 	 y ~ x 
#>   Exposure a (Propensity model: logistic regression):
#> 	 a ~ x 
#> 
#>                   Estimate Std.Err    2.5%   97.5%    P-value
#>  a=0               0.48506 0.02626  0.4336  0.5365  3.458e-76
#>  a=1               0.67794 0.02225  0.6343  0.7215 6.005e-204
#> Outcome model:                                               
#>  (Intercept)       0.44427 0.07306  0.3011  0.5875  1.196e-09
#>  x                 1.06929 0.08537  0.9020  1.2366  5.408e-36
#> Propensity model:                                            
#>  (Intercept)       0.06214 0.09258 -0.1193  0.2436  5.021e-01
#>  x                -0.92905 0.15311 -1.2291 -0.6289  1.297e-09
#> 
#> Average Causal Effect (constrast: 'a=0' vs. 'a=1'):
#> 
#>    Estimate Std.Err    2.5%   97.5%   P-value
#> RR   0.7155 0.04356  0.6301  0.8009 1.259e-60
#> OR   0.4475 0.06268  0.3246  0.5703 9.383e-13
#> RD  -0.1929 0.03295 -0.2575 -0.1283 4.791e-09
#> 
#> 
#> ace> # Multiple treatments
#> ace> m <- lvm(y ~ a+x, a~x)
#> 
#> ace> distribution(m,~ y) <- binomial.lvm()
#> 
#> ace> m <- ordinal(m, K=4, ~a)
#> 
#> ace> transform(m, ~a) <- factor
#> 
#> ace> d <- sim(m,1e4)
#> 
#> ace> (a <- ace(y~a|a*x|x, data=d))
#>     Estimate  Std.Err   2.5%  97.5%    P-value
#> a=0   0.2502 0.026118 0.1990 0.3013  9.928e-22
#> a=1   0.3511 0.018925 0.3140 0.3882  7.815e-77
#> a=2   0.4329 0.015141 0.4032 0.4626 8.579e-180
#> a=3   0.6211 0.008511 0.6044 0.6378  0.000e+00
#> 
#> ace> # Comparison with randomized experiment
#> ace> m0 <- cancel(m, a~x)
#> 
#> ace> d0 <- sim(m0,2e5)
#> 
#> ace> lm(y~a-1,d0)
#> 
#> Call:
#> lm(formula = y ~ a - 1, data = d0)
#> 
#> Coefficients:
#>     a0      a1      a2      a3  
#> 0.2291  0.3411  0.4133  0.6211  
#> 
#> 
#> ace> # Choosing a different contrast for the association measures
#> ace> summary(a, contrast=c(2,4))
#> 
#> Augmented Inverse Probability Weighting estimator
#>   Response y (Outcome model: logistic regression):
#> 	 y ~ a * x 
#>   Exposure a (Propensity model: logistic regression):
#> 	 a ~ x 
#> 
#>     Estimate  Std.Err   2.5%  97.5%    P-value
#> a=0   0.2502 0.026118 0.1990 0.3013  9.928e-22
#> a=1   0.3511 0.018925 0.3140 0.3882  7.815e-77
#> a=2   0.4329 0.015141 0.4032 0.4626 8.579e-180
#> a=3   0.6211 0.008511 0.6044 0.6378  0.000e+00
#> 
#> Average Causal Effect (constrast: 'a=1' vs. 'a=3'):
#> 
#>    Estimate Std.Err    2.5%   97.5%   P-value
#> RR   0.5653 0.03063  0.5053  0.6253 4.620e-76
#> OR   0.3301 0.02871  0.2738  0.3863 1.391e-30
#> RD  -0.2700 0.01991 -0.3090 -0.2310 6.562e-42
#>