Targeted inference

Methods for targeted and semiparametric inference including augmented inverse probability weighted estimators for missing data and causal inference.

Author

Klaus K. Holst (Maintainer) klaus@holst.it

Examples

example(riskreg)
#> 
#> riskrg> m <- lvm(a[-2] ~ x,
#> riskrg+          z ~ 1,
#> riskrg+          lp.target[1] ~ 1,
#> riskrg+          lp.nuisance[-1] ~ 2*x)
#> 
#> riskrg> distribution(m,~a) <- binomial.lvm("logit")
#> 
#> riskrg> m <- binomial.rr(m, "y","a","lp.target","lp.nuisance")
#> 
#> riskrg> d <- sim(m,5e2,seed=1)
#> 
#> riskrg> I <- model.matrix(~1, d)
#> 
#> riskrg> X <- model.matrix(~1+x, d)
#> 
#> riskrg> with(d, riskreg_mle(y, a, I, X, type="rr"))
#>                          Estimate Std.Err    2.5%   97.5%   P-value
#> (Intercept)                0.9453  0.1150  0.7198  1.1707 2.092e-16
#> odds-product:(Intercept)  -0.8068  0.3031 -1.4009 -0.2126 7.783e-03
#> odds-product:x             2.3296  0.3914  1.5624  3.0969 2.659e-09
#> 
#> riskrg> with(d, riskreg_fit(y, a, nuisance=X, propensity=I, type="rr"))
#>             Estimate Std.Err   2.5% 97.5%   P-value
#> (Intercept)   0.9469   0.113 0.7254 1.168 5.418e-17
#> 
#> riskrg> riskreg(y ~ a | 1, nuisance=~x ,  data=d, type="rr")
#>             Estimate Std.Err   2.5% 97.5%   P-value
#> (Intercept)   0.9469   0.113 0.7254 1.168 5.418e-17
#> 
#> riskrg> ## Model with same design matrix for nuisance and propensity model:
#> riskrg> with(d, riskreg_fit(y, a, nuisance=X, type="rr"))
#>             Estimate Std.Err   2.5% 97.5%   P-value
#> (Intercept)   0.9402  0.1173 0.7104  1.17 1.082e-15
#> 
#> riskrg> ## a <- riskreg(y ~ a, target=~z, nuisance=~x,  propensity=~x, data=d, type="rr")
#> riskrg> a <- riskreg(y ~ a | z, nuisance=~x,  propensity=~x, data=d, type="rr")
#> 
#> riskrg> a
#>             Estimate Std.Err    2.5%     97.5%   P-value
#> (Intercept)   0.9797  0.1212  0.7421  1.217281 6.366e-16
#> z            -0.2218  0.1115 -0.4404 -0.003198 4.674e-02
#> 
#> riskrg> riskreg(y ~ a, nuisance=~x,  data=d, type="rr", mle=TRUE)
#>                          Estimate Std.Err    2.5%   97.5%   P-value
#> (Intercept)                0.9453  0.1150  0.7198  1.1707 2.092e-16
#> odds-product:(Intercept)  -0.8068  0.3031 -1.4009 -0.2126 7.783e-03
#> odds-product:x             2.3296  0.3914  1.5624  3.0969 2.659e-09
example(cate)
#> 
#> cate> sim1 <- function(n=1e4,
#> cate+                  seed=NULL,
#> cate+                  return_model=FALSE, ...) {
#> cate+ suppressPackageStartupMessages(require("lava"))
#> cate+ if (!is.null(seed)) set.seed(seed)
#> cate+ m <- lava::lvm()
#> cate+ regression(m, ~a) <- function(z1,z2,z3,z4,z5)
#> cate+          cos(z1)+sin(z1*z2)+z3+z4+z5^2
#> cate+ regression(m, ~u) <- function(a,z1,z2,z3,z4,z5)
#> cate+         (z1+z2+z3)*a + z1+z2+z3 + a
#> cate+ distribution(m, ~a) <- binomial.lvm()
#> cate+ if (return_model) return(m)
#> cate+ lava::sim(m, n, p=par)
#> cate+ }
#> 
#> cate> d <- sim1(200)
#> 
#> cate> if (require("SuperLearner",quietly=TRUE)) {
#> cate+   e <- cate(a ~ z1+z2+z3, response=u~., data=d)
#> cate+   e
#> cate+ }
#>             Estimate Std.Err    2.5% 97.5%   P-value
#> (Intercept)   0.9841  0.2682 0.45842 1.510 2.433e-04
#> z1            1.3639  0.3206 0.73544 1.992 2.103e-05
#> z2            0.8310  0.1769 0.48425 1.178 2.635e-06
#> z3            0.5569  0.2747 0.01853 1.095 4.262e-02
example(ate)
#> 
#> ate> m <- lvm(y ~ a+x, a~x)
#> 
#> ate> distribution(m,~ a+y) <- binomial.lvm()
#> 
#> ate> d <- sim(m,1e3,seed=1)
#> 
#> ate> a <- ate(y ~ a, nuisance=~x, data=d)
#> 
#> ate> summary(a)
#> 
#> Augmented Inverse Probability Weighting estimator
#>   Response y (Outcome model: logistic regression):
#> 	 y ~ x + a 
#>   Exposure a (Propensity model: logistic regression):
#> 	 a ~ x 
#> 
#>                   Estimate Std.Err     2.5%   97.5%    P-value
#>  a=0               0.48496 0.02311  0.43967  0.5302  8.607e-98
#>  a=1               0.67788 0.02240  0.63397  0.7218 4.249e-201
#> Outcome model:                                                
#>  (Intercept)      -0.02229 0.10163 -0.22149  0.1769  8.264e-01
#>  x                 0.90226 0.08857  0.72866  1.0759  2.276e-24
#>  a                 1.01052 0.15195  0.71270  1.3083  2.926e-11
#> Propensity model:                                             
#>  (Intercept)       0.06214 0.06914 -0.07337  0.1977  3.688e-01
#>  x                -0.92905 0.08035 -1.08653 -0.7716  6.408e-31
#> 
#> Average Treatment Effect (constrast: 'a=0' vs. 'a=1'):
#> 
#>    Estimate Std.Err    2.5%   97.5%   P-value
#> RR   0.7154 0.04004  0.6369  0.7939 2.114e-71
#> OR   0.4474 0.05952  0.3308  0.5641 5.611e-14
#> RD  -0.1929 0.03099 -0.2537 -0.1322 4.785e-10
#> 
#> 
#> ate> # Multiple treatments
#> ate> m <- lvm(y ~ a+x, a~x)
#> 
#> ate> distribution(m,~ y) <- binomial.lvm()
#> 
#> ate> m <- ordinal(m, K=4, ~a)
#> 
#> ate> transform(m, ~a) <- factor
#> 
#> ate> d <- sim(m, 1e4, seed=1)
#> 
#> ate> (a <- ate(y~a|a*x|x, data=d))
#>     Estimate  Std.Err   2.5%  97.5%    P-value
#> a=0   0.2406 0.018291 0.2047 0.2764  1.638e-39
#> a=1   0.3638 0.017241 0.3300 0.3976  7.816e-99
#> a=2   0.4003 0.015525 0.3699 0.4307 1.286e-146
#> a=3   0.6054 0.007469 0.5908 0.6201  0.000e+00
#> 
#> ate> # Comparison with randomized experiment
#> ate> m0 <- cancel(m, a~x)
#> 
#> ate> d0 <- sim(m0,2e5)
#> 
#> ate> lm(y~a-1,d0)
#> 
#> Call:
#> lm(formula = y ~ a - 1, data = d0)
#> 
#> Coefficients:
#>     a0      a1      a2      a3  
#> 0.2232  0.3393  0.4093  0.6208  
#> 
#> 
#> ate> # Choosing a different contrast for the association measures
#> ate> summary(a, contrast=c(2,4))
#> 
#> Augmented Inverse Probability Weighting estimator
#>   Response y (Outcome model: logistic regression):
#> 	 y ~ a * x 
#>   Exposure a (Propensity model: logistic regression):
#> 	 a ~ x 
#> 
#>     Estimate  Std.Err   2.5%  97.5%    P-value
#> a=0   0.2406 0.018291 0.2047 0.2764  1.638e-39
#> a=1   0.3638 0.017241 0.3300 0.3976  7.816e-99
#> a=2   0.4003 0.015525 0.3699 0.4307 1.286e-146
#> a=3   0.6054 0.007469 0.5908 0.6201  0.000e+00
#> 
#> Average Treatment Effect (constrast: 'a=1' vs. 'a=3'):
#> 
#>    Estimate Std.Err    2.5%   97.5%   P-value
#> RR   0.6009 0.02911  0.5438  0.6579 1.240e-94
#> OR   0.3726 0.02964  0.3146  0.4307 2.964e-36
#> RD  -0.2416 0.01849 -0.2779 -0.2054 4.999e-39
#> 
example(calibration)
#> 
#> clbrtn> sim1 <- function(n, beta=c(-3, rep(.5,10)), rho=.5) {
#> clbrtn+  p <- length(beta)-1
#> clbrtn+  xx <- lava::rmvn0(n,sigma=diag(nrow=p)*(1-rho)+rho)
#> clbrtn+  y <- rbinom(n, 1, lava::expit(cbind(1,xx)%*%beta))
#> clbrtn+  d <- data.frame(y=y, xx)
#> clbrtn+  names(d) <- c("y",paste0("x",1:p))
#> clbrtn+  return(d)
#> clbrtn+ }
#> 
#> clbrtn> set.seed(1)
#> 
#> clbrtn> beta <- c(-2,rep(1,10))
#> 
#> clbrtn> d <- sim1(1e4, beta=beta)
#> 
#> clbrtn> a1 <- NB(y ~ ., data=d)
#> 
#> clbrtn> a2 <- glm(y ~ ., data=d, family=binomial)
#> Warning: glm.fit: fitted probabilities numerically 0 or 1 occurred
#> 
#> clbrtn> ## a3 <- randomForest(factor(y) ~ ., data=d, family=binomial)
#> clbrtn> 
#> clbrtn> d0 <- sim1(1e5, beta=beta)
#> 
#> clbrtn> p1 <- predict(a1, newdata=d0)
#> 
#> clbrtn> p2 <- predict(a2, newdata=d0, type="response")
#> 
#> clbrtn> ## p3 <- predict(a3, newdata=d0, type="prob")
#> clbrtn> 
#> clbrtn> c2 <- calibration(p2, d0$y, method="isotonic")
#> 
#> clbrtn> c1 <- calibration(p1, d0$y, breaks=100)
#> 
#> clbrtn> if (interactive()) {
#> clbrtn+   plot(c1)
#> clbrtn+   plot(c2,col="red",add=TRUE)
#> clbrtn+   abline(a=0,b=1)##'
#> clbrtn+   with(c1$xy[[1]], points(pred,freq,type="b", col="red"))
#> clbrtn+ }
#> 
#> clbrtn> set.seed(1)
#> 
#> clbrtn> beta <- c(-2,rep(1,10))
#> 
#> clbrtn> dd <- lava::csplit(sim1(6e4, beta=beta), k=3)
#> 
#> clbrtn> mod <- NB(y ~ ., data=dd[[1]])
#> 
#> clbrtn> p1 <- predict(mod, newdata=dd[[2]])
#> 
#> clbrtn> cal <- calibration(p1, dd[[2]]$y)
#> 
#> clbrtn> p2 <- predict(mod, newdata=dd[[3]])
#> 
#> clbrtn> pp <- predict(c1, p2)
#> 
#> clbrtn> cc <- calibration(pp, dd[[3]]$y)
#> 
#> clbrtn> if (interactive()) {##'
#> clbrtn+   plot(cal)
#> clbrtn+   plot(cc, add=TRUE, col="blue")
#> clbrtn+ }